+++ packages/plugin-commons/web/src/Components/ColorPicker/ColorPicker.jsx
@@ -163,7 +163,7 @@ class ColorPicker extends PureComponent {
     } = this;
     const { t, isMobile, theme, children } = this.props;
     return (
-      <div className={classNames(styles.colorPicker)} tabIndex={0}>
+      <div className={styles.colorPicker} tabIndex={0}>
         {this.state.isCustomColorPickerOpened
           ? this.props.onCustomPickerToggle({
               color: this.state.color,
@@ -175,9 +175,9 @@ class ColorPicker extends PureComponent {
               theme,
             })
           : children({
-              renderAddColorButton,
               renderPalette,
               renderUserColors,
+              renderAddColorButton,
               renderResetColorButton,
               mergedStyles: styles,
             })}